internal: move pkg/system to internal
