+++ oci/casext/blobcompress/gzip.go
@@ -23,8 +23,8 @@ import (
 	"io"
 	"runtime"
 
+	"github.com/opencontainers/umoci/internal/system"
 	"github.com/opencontainers/umoci/oci/casext/mediatype"
-	"github.com/opencontainers/umoci/pkg/system"
 
 	"github.com/apex/log"
 	gzip "github.com/klauspost/pgzip"
