+++ oci/layer/tar_generate.go
@@ -31,8 +31,8 @@ import (
 	"golang.org/x/sys/unix"
 
 	"github.com/opencontainers/umoci/internal/funchelpers"
+	"github.com/opencontainers/umoci/internal/system"
 	"github.com/opencontainers/umoci/pkg/fseval"
-	"github.com/opencontainers/umoci/pkg/system"
 )
 
 // tarGenerator is a helper for generating layer diff tars. It should be noted
