+++ pkg/fseval/fseval_default.go
@@ -27,7 +27,7 @@ import (
 	"golang.org/x/sys/unix"
 
 	"github.com/opencontainers/umoci/internal/funchelpers"
-	"github.com/opencontainers/umoci/pkg/system"
+	"github.com/opencontainers/umoci/internal/system"
 )
 
 // Default is the "identity" form of FsEval. In particular, it does not do any
