+++ pkg/unpriv/unpriv.go
@@ -44,7 +44,7 @@ import (
 	"golang.org/x/sys/unix"
 
 	"github.com/opencontainers/umoci/internal/funchelpers"
-	"github.com/opencontainers/umoci/pkg/system"
+	"github.com/opencontainers/umoci/internal/system"
 )
 
 // fiRestore restores the state given by an os.FileInfo instance at the given
